Adding support to WhatsApp events on Business SDK
Summary:
This diff adds support for WhatsApp events on the Business SDK. The changes include adding two new fields to the UserData object and a new field to the ServerEvent object. The changes are made in the user-data.js and server-event.js files. The summary of the changes is as follows:

- A new field ctwa_clid is added to the UserData object to store the ID of a conversation that was started on WhatsApp.
- A new field page_id is added to the UserData object to store the ID of the page that the ad is associated with.
- A new field messaging_channel is added to the ServerEvent object to store the messaging channel of the event (e.g. WhatsApp).

Reviewed By: stcheng

Differential Revision: D53942902

fbshipit-source-id: cbada5a34983e9e4fa067b266a1305790c4bd684