Create SDK for creating ThirdPartyAccounts
Summary:
# Background
In order to simplify the usages for the two tier business model, we are creating SDK classes that abstract away complexity of setup in an effort to reduce onboarding friction. This contains both the classes as well as examples on how to use them.

This SDK Library helps create a third party account under this tech provider's business. This account is used to represent a third party user on the tech provider's platform.

 The account allows the tech provider's application to:
1. Create Assets (Ad Accounts, Pages, Pixels, etc) that are owned by the tech provider.
2. Access Assets (Pages, Pixels, etc) from a users facebook account or business account.
3. Allocate a portion of the tech provider's Credit Line to the account.
4. Use one access token to manage both created and shared assets.

Further, it allows the tech provider to decide which assets are shared from an existing facebook user, and which assets are created and owned by the tech provider. This allows the tech provider to provide monetization experiences for the third party user using a higher level abstraction without worrying about access token management.

# Features
This diff implements the core functionalities:
- creating a third party account and associated ad account
- creating and sharing a line of credit with the third
- fetching a particular child business account
- fetching the list of child business accounts
- getting the Facebook page of a child business
- access token management is abstracted away as there are 3 tokens involved

Differential Revision: D73810049

fbshipit-source-id: a703883d2b0c06e814b9f77a706a5f7814d905a7